refactor: separate farm and plot productivity calculation
