+++ src/main/java/com/aegro/api/service/PlotService.java
@@ -21,7 +21,7 @@ public interface PlotService {
 
 	void updateProductivityByPlotId(Long idPlot);
 
-	Double getProductivityByIdPlot(Long idPlot);
+//	Double getProductivityByIdPlot(Long idPlot);
 
 	Plot createPlotInFarmId(Plot plot, Long idFarm);
