+++ src/main/java/com/aegro/api/service/ProductivityFarm.java
@@ -1,0 +1,17 @@
+package com.aegro.api.service;
+
+/**
+ * @author Yohan Silva
+ */
+
+public interface ProductivityFarm {
+	
+	void updateProductivityFarm(Long idFarm);
+	
+	Double getTotalAreaByFarmId(Long idFarm);
+	
+	Double getProductionFarmById(Long idFarm);
+	
+	Double getFarmProductivityById(Long idFarm);
+	
+}
