+++ src/main/java/com/aegro/api/service/ProductivityPlot.java
@@ -1,0 +1,14 @@
+package com.aegro.api.service;
+
+/**
+ * @author Yohan Silva
+ */
+public interface ProductivityPlot {
+
+	Double getProductivityByIdPlot(Long idPlot);
+
+	void updateProductivityByPlotId(Long idPlot);
+	
+	Double calculatePlotProductivity(Long idPlot);
+	
+}
