--- src/main/resources/data.sql
@@ -1,36 +1,0 @@
-INSERT INTO FARM(name_Farm) VALUES('Fazenda SaoPaulo');
-INSERT INTO FARM(name_Farm) VALUES('Fazenda Curitiba');
-INSERT INTO FARM(name_Farm) VALUES('Fazenda Salvador');
-INSERT INTO FARM(name_Farm) VALUES('Fazenda PortoAlegre');
-INSERT INTO FARM(name_Farm) VALUES('Fazenda Natal');
-
-
-INSERT INTO PLOT(name_Plot, plot_Area_In_Hectare, farm_id) VALUES('Talhao Roxo', '200', 1);
-INSERT INTO PLOT(name_Plot, plot_Area_In_Hectare, farm_id) VALUES('Talhao Azul', '10', 2);
-INSERT INTO PLOT(name_Plot, plot_Area_In_Hectare, farm_id) VALUES('Talhao Amarelo', '245', 2);
-INSERT INTO PLOT(name_Plot, plot_Area_In_Hectare, farm_id) VALUES('Talhao Cinza', '203', 2);
-INSERT INTO PLOT(name_Plot, plot_Area_In_Hectare, farm_id) VALUES('Talhao Purpura', '211', 3);
-INSERT INTO PLOT(name_Plot, plot_Area_In_Hectare, farm_id) VALUES('Talhao Lilas', '500', 4);
-INSERT INTO PLOT(name_Plot, plot_Area_In_Hectare, farm_id) VALUES('Talhao Azul Marinho', '1200', 1);
-INSERT INTO PLOT(name_Plot, plot_Area_In_Hectare, farm_id) VALUES('Talhao Beje', '235', 4);
-
-
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('230', 4);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('900', 3);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('345', 1);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('1000', 3);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('3000', 5);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('50', 5);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('400', 2);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('34', 6);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('1200', 7);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('500', 4);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('300', 3);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('730', 1);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('10', 3);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('902', 5);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('357', 5);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('250', 2);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('401', 6);
-INSERT INTO PRODUCTION(amount, plot_id) VALUES('120', 7);
-
